Move class of command Framework:init
